refactor: re-structuring the components
