+++ src/components/Home/CommunicateSection/__test__/CommunicateThroughSection.test.js
@@ -1,0 +1,8 @@
+import renderer from "react-test-renderer";
+
+import CommunicateThroughSection from "../CommunicateThroughSection";
+
+it("renders correctly", () => {
+    const tree = renderer.create(<CommunicateThroughSection />).toJSON();
+    expect(tree).toMatchSnapshot();
+});
