+++ src/components/Home/HeroSection.jsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 
-import Button from "./ui/Button";
+import Button from "../ui/Button";
 
 export default function Hero(prop) {
     const { t } = prop;
