+++ src/components/Home/RecentBlogSection/__test__/RecentBlogSection.test.js
@@ -1,0 +1,7 @@
+import renderer from "react-test-renderer";
+
+import RecentBlogSection from "../RecentBlogSection";
+it("renders correctly", () => {
+    const tree = renderer.create(<RecentBlogSection />).toJSON();
+    expect(tree).toMatchSnapshot();
+});
