+++ src/components/Home/TicketSection/TicketItem.jsx
