+++ src/components/Home/TicketSection/__test__/TicketItem.test.js
