+++ src/components/Home/TicketSection/__test__/TicketSection.test.js
