--- src/components/communicate/__test__/CommunicateThrough.test.js
@@ -1,8 +1,0 @@
-import renderer from "react-test-renderer";
-
-import CommunicateThrough from "../CommunicateThrough";
-
-it("renders correctly", () => {
-    const tree = renderer.create(<CommunicateThrough />).toJSON();
-    expect(tree).toMatchSnapshot();
-});
